pyproject.toml
[project]
name = "edinet2dataset"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "speed1313", email = "speedtry13@icloud.com" }
]
dependencies = [
    "python-dotenv>=1.0.1",
    "tqdm>=4.67.1",
    "requests>=2.32.3",
    "polars>=1.26.0",
    "loguru>=0.7.3",
    "anthropic>=0.50.0",
    "backoff>=2.2.1",
    "pdfminer.six>=20250416",
]
readme = "README.md"
requires-python = ">= 3.10"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"



[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/edinet2dataset"]

[dependency-groups]
dev = [
    "datasets>=3.5.0",
    "japanize-matplotlib>=1.1.3",
    "matplotlib-fontja>=1.0.1",
    "matplotlib>=3.10.1",
    "pytest>=8.3.5",
    "ruff>=0.11.4",
    "scikit-learn>=1.6.1",
    "transformers>=4.51.1",
    "markdownify>=1.1.0",
    "trafilatura>=2.0.0",
    "xbrr>=0.2.7.5",
]
